requirements.txt
videosdk
python-dotenv
